requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
Appium-Python-Client==1.2.0
pytest==6.2.2
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.1
python-dateutil==2.8.1
pywin32==227
pywin32-ctypes==0.2.0
pywinauto==0.6.8
selenium==3.141.0